feat: Fields pay for each site managment.
Add routes and methods for save a new field for a site.

delete a field for a site.

Validate the const to show in eachs site.

Clean the seeder Fieldspaysite to save just the needed information.

Modify the view for add the correct functionality

Add the routes for each site type